Fixed int dividing

+++ TagsCloudContainer/CloudVisualizers/BitmapMakers/DefaultBitmapMaker.cs
@@ -65,7 +65,7 @@ namespace TagsCloudContainer.CloudVisualizers.BitmapMakers
             if (xDifference < yDifference)
                 (xRatio, yRatio) = GetVerticalLayoutRatio(settings, maxX, xDifference, yDifference);
             if (settings.Width == settings.Height)
-                (xRatio, yRatio) = (settings.Width / xDifference, settings.Height / yDifference);
+                (xRatio, yRatio) = ((float)settings.Width / xDifference, (float)settings.Height / yDifference);
 
             return (xRatio, yRatio);
         }
